Use a WeakKeyDictionary to store session => status map
Stale entries for old sessions could be left in the `id(session) =>
status` map if an application failed to end a transaction after a
session event, from the set used by `zope.sqlalchemy.register`, caused
the session to be joined to the transaction.

While such issues often indicate an error on the part of the
application, this can lead to difficult-to-debug issues because IDs can
be re-used in the lifetime of a Python process, so a stale entry for an
old Session may affect what happens to a future, unrelated Session.

Using a weak dict of `session => state` avoids this hazard.